Added GUI element to select distortion mode. Replaced old intrinsics calibration.
