+++ pupil_src/shared_modules/calibration_routines/__init__.py
@@ -16,7 +16,6 @@ from . manual_marker_calibration import Manual_Marker_Calibration
 from . single_marker_calibration import Single_Marker_Calibration
 from . natural_features_calibration import Natural_Features_Calibration
 from . camera_intrinsics_estimation import Camera_Intrinsics_Estimation
-from . camera_intrinsics_estimation_fisheye import Camera_Intrinsics_Estimation_Fisheye
 from . adjust_calibration import Adjust_Calibration
 from . accuracy_test import Accuracy_Test
 from . hmd_calibration import HMD_Calibration, HMD_Calibration_3D
@@ -28,7 +27,6 @@ calibration_plugins = [Screen_Marker_Calibration,
                        Single_Marker_Calibration,
                        Natural_Features_Calibration,
                        Camera_Intrinsics_Estimation,
-                        Camera_Intrinsics_Estimation_Fisheye,
                        Accuracy_Test,
                        Adjust_Calibration,
                        HMD_Calibration,
